+++ activestorage/app/controllers/active_storage/disk_controller.rb
@@ -33,7 +33,7 @@ class ActiveStorage::DiskController < ActiveStorage::BaseController
 
   private
     def disk_service
-      ActiveStorage::Blob.service
+      ActiveStorage.blob_class.service
     end