+++ activestorage/lib/active_storage/attached/changes/delete_many.rb
@@ -9,11 +9,11 @@ module ActiveStorage
     end
 
     def attachments
-      ActiveStorage::Attachment.none
+      ActiveStorage.attachment_class.none
     end
 
     def blobs
-      ActiveStorage::Blob.none
+      ActiveStorage.blob_class.none
     end
 
     def save