+++ activestorage/lib/active_storage/engine.rb
@@ -73,6 +73,8 @@ module ActiveStorage
         ActiveStorage.analyzers         = app.config.active_storage.analyzers || []
         ActiveStorage.paths             = app.config.active_storage.paths || {}
         ActiveStorage.routes_prefix     = app.config.active_storage.routes_prefix || "/rails/active_storage"
+        ActiveStorage.blob_class        = app.config.active_storage.blob_class || "ActiveStorage::Blob"
+        ActiveStorage.attachment_class  = app.config.active_storage.attachment_class || "ActiveStorage::Attachment"
 
         ActiveStorage.variable_content_types = app.config.active_storage.variable_content_types || []
         ActiveStorage.content_types_to_serve_as_binary = app.config.active_storage.content_types_to_serve_as_binary || []